requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
pytz>=2023.3
tweepy>=4.14.0
pandas>=2.1.0
lxml>=4.9.3
ics>=0.7.2
pyshorteners>=1.0.1 